+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Add support for `default` key in `launch.toml` in `Process` struct
 - Support the new `buildpack.toml` fields `description`, `keywords` and `licenses`
 - Set a minumim required Rust version of 1.56 and switch to the 2021 Rust edition
 - Stack id in `buildpack.toml` can now be `*` indicating "any" stack